+++ src/lib/types/availability/availability.types.ts
@@ -1,0 +1,5 @@
+export interface Availability {
+    isAvailable: boolean;
+    reservations: number;
+    available: number;
+}