+++ src/lib/types/availability/index.ts
@@ -1,0 +1,3 @@
+export * from './availability.types';
+
+export * from './update-params.type';