+++ src/lib/types/availability/update-params.type.ts
@@ -1,0 +1,4 @@
+export interface UpdateParams {
+    reservations: number;
+    available: number;
+}