+++ src/lib/types/reservation/index.ts
@@ -1,0 +1,3 @@
+export * from './reservation-operation.type';
+
+export * from './create-reservation.type';