+++ src/lib/types/reservation/reservation-operation.type.ts
@@ -1,0 +1,5 @@
+export enum ReservationOperation {
+    ADD = 'ADD',
+    SUBTRACT = 'SUBTRACT'
+}
+